Rename HGMO -> HgRev and stop tying 'json-pushes' data to an instance
Since instances are tied to a branch and revision, it makes little sense
to store push data on them.

This pulls all the 'json-pushes' data out into staticmethods and a class
level cache. It also adds a convenience method for loading data from a
single push which uses the cache.