Only try to gather groups from completed tasks
This also fixes a bug in the taskcluster data source where
we normalize 'exception' to 'completed'.

In treeherder exception means the task ran and had a problem. But in
taskcluster land, an exception mean the task never even ran. E.g, a task
can be in this state if its dependency failed and then hit an expiry.